Start-Up scene stack loading
Changed start-up scene controller to take scene stack strings as input to specify how to start the game instead of a single scene's name. This means that the game can be started using a scene stack instead of a single scene. This will mostly be useful for testing as the main game system will probably work differently to this