Saving and loading functionality
Added functionality for saving and loading of saved games using the save menu allowing for 5 different saved game files (a.k.a. slots) that can be overwritten and loaded anytime that the save menu can be opened.
Trying to load an empty save currently does nothing but will later be used for starting new games. Also, when the game is launched, the starting menu will later be changed to allowing the player to select a save slot to load and also allowing them to start a new game